+++ spyne/interface/xml_schema/_base.py
@@ -20,6 +20,7 @@
 import logging
 logger = logging.getLogger('spyne.interface.xml_schema')
 
+import os
 import shutil
 import tempfile
 
@@ -210,22 +211,23 @@ class XmlSchema(InterfaceDocumentBase):
         try:
             # serialize nodes to files
             for k, v in self.schema_dict.items():
-                file_name = '%s/%s.xsd' % (tmp_dir_name, k)
+                file_name = os.path.join(tmp_dir_name, "%s.xsd" % k)
                 with open(file_name, 'wb') as f:
                     etree.ElementTree(v).write(f, pretty_print=True)
 
                 logger.debug("writing %r for ns %s" %
                              (file_name, self.interface.nsmap[k]))
 
-            with open('%s/%s.xsd' % (tmp_dir_name, pref_tns), 'r') as f:
+            with open(os.path.join(tmp_dir_name, "%s.xsd" % pref_tns), 'r') as f:
                 try:
                     self.validation_schema = etree.XMLSchema(etree.parse(f))
 
                 except Exception:
                     f.seek(0)
-                    logger.error("This is a Spyne error. Please seek support "
-                                 "with a minimal test case that reproduces "
-                                 "this error.")
+                    logger.error("This could be a Spyne error. Unless you're "
+                                 "sure the reason for this error is outside "
+                                 "Spyne, please open a new issue with a "
+                                 "minimal test case that reproduces it.")
                     raise
 
             shutil.rmtree(tmp_dir_name)