+++ spyne/util/web.py
@@ -322,17 +322,26 @@ def log_repr(obj, cls=None, given_len=None, parent=None, from_array=False, tags=
                 retval.append("(...)")
                 break
 
-            if t.Attributes.logged:
-                v = getattr(obj, k, None)
-                # HACK!: sometimes non-db attributes restored from database don't
-                # get properly reinitialized.
-                if isclass(v) and issubclass(v, ModelBase):
-                    continue
+            if not t.Attributes.logged:
+                continue
 
-                if v is not None:
-                    retval.append("%s=%s" % (k, log_repr(v, t, parent=k,
-                                                                    tags=tags)))
-                    i += 1
+            try:
+                v = getattr(obj, k, None)
+            except (AttributeError, KeyError):
+                v = None
+
+            # HACK!: sometimes non-db attributes restored from database don't
+            # get properly reinitialized.
+            if isclass(v) and issubclass(v, ModelBase):
+                continue
+
+            polymap = t.Attributes.polymap
+            if polymap is not None:
+                t = polymap.get(v.__class__, t)
+
+            if v is not None:
+                retval.append("%s=%s" % (k, log_repr(v, t, parent=k, tags=tags)))
+                i += 1
 
         return "%s(%s)" % (cls.get_type_name(), ', '.join(retval))